+++ src/main/java/dungeon/models/Position.java
@@ -1,20 +1,20 @@
 package dungeon.models;
 
 public class Position {
-  private final float x;
+  private final int x;
 
-  private final float y;
+  private final int y;
 
-  public Position (float x, float y) {
+  public Position (int x, int y) {
     this.x = x;
     this.y = y;
   }
 
-  public float getX () {
+  public int getX () {
     return this.x;
   }
 
-  public float getY () {
+  public int getY () {
     return this.y;
   }
 }
